Add central styles/formatting record storage
This is an initial commit to create centralized repozitory for formatting
records. The XLStyleKey/XLStyleValue are a final formatting records that
specify all properties of format.

That is not how styles are represented or used in workbook. Formatting
record contains only some selected attributes, the rest is determined
by other formatting records in a hiearchy

* default formatting
* workbook formatting (aka normal style)
* row/column formatting
* cell formatting

This final-format-representation has many problems when it comes to
modifying styles. Named styles are also missing. Current code also drags
through many properties variousl elements of formatting instead of keeping
it in one place.

In short term the loading code will use the repository to determine cell
styles, in medium term it should be used longer term