Make CodeGen simple types stringly typed (#2679)
CodeGen originally used generic arguments for simple types (e.g.
AddSimpleTypeRequired<int>). That was a wrong way to do it, because CodeGen
doesn't have reference to the ClosedXML project. Without reference to the
ClosedXML project, it wasn't possible to define simple type enum mapping.

This PR changes the XML simple type <-> C# type from strongly typed to
stringly typed. Stringly typing allows use of types that are not referenced,
including enums

The added styles reader stub is going to be playground to improve the
CodeGen hooks.