Add default format and format fallbacks
This is a WIP (not used anywhere) code to replace the XLStyle and
XLStylized.

The goal is to get the IXLStyle getters. They need to return a value
even if cell or row doesn't have a format (or in case of row don't even
exist yet). To do that, they go up the hierarchy until they find
element with a format.

Once they have the format, some values might not be specified. Although
it's very likely Excel has everything materialized, I am going with the
missing values + default. Mostly because technically that is part of
the spec and moving from missing+default to everything specified is far
simpler than the other way.

The style of a format is not yet used, or style or apply* flags. That
will be done later.